RegisterApp/requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
SQLAlchemy>=2.0
psycopg2-binary>=2.9
openpyxl>=3.1
gunicorn>=21.2
werkzeug>=3.0